refactor(forge): Split forge contracts
The forge package owns forge-neutral contracts, but forge.go combined
registry matching, repository interfaces, change models, merge behavior,
mergeability vocabulary, comment models, and check models in one file. Split
those concepts into responsibility-named files so changes to one contract area
do not require navigating every forge-neutral model.

This is a package-local file organization change. Public symbols, interface
method sets, serialization behavior, and concrete forge implementations remain
unchanged.
